--- a/build.zig
+++ b/build.zig
@@ -21,16 +21,14 @@ pub fn build(b: *std.Build) void {
 }
 
 fn setupLibrary(b: *std.Build, target: std.Build.ResolvedTarget, optimize: std.builtin.OptimizeMode) *std.Build.Step.Compile {
-    const module = b.createModule(.{
-        .root_source_file = b.path("zbench.zig"),
-        .target = target,
-        .optimize = optimize,
-    });
-
     const lib = b.addLibrary(.{
         .linkage = .static,
         .name = "zbench",
-        .root_module = module,
+        .root_module = b.createModule(.{
+            .root_source_file = b.path("zbench.zig"),
+            .target = target,
+            .optimize = optimize,
+        }),
         .version = version,
     });
 
@@ -50,14 +48,13 @@ fn setupTesting(b: *std.Build, target: std.Build.ResolvedTarget, optimize: std.b
 
     const test_step = b.step("test", "Run library tests");
     for (test_files) |test_file| {
-        const test_module = b.createModule(.{
-            .root_source_file = .{ .src_path = .{ .owner = b, .sub_path = test_file.path } },
-            .target = target,
-            .optimize = optimize,
-        });
         const _test = b.addTest(.{
             .name = test_file.name,
-            .root_module = test_module,
+            .root_module = b.createModule(.{
+                .root_source_file = .{ .src_path = .{ .owner = b, .sub_path = test_file.path } },
+                .target = target,
+                .optimize = optimize,
+            }),
         });
         const run_test = b.addRunArtifact(_test);
         test_step.dependOn(&run_test.step);
@@ -82,14 +79,13 @@ fn setupExamples(b: *std.Build, target: std.Build.ResolvedTarget, optimize: std.
     };
 
     for (example_names) |example_name| {
-        const example_module = b.createModule(.{
-            .root_source_file = .{ .src_path = .{ .owner = b, .sub_path = b.fmt("examples/{s}.zig", .{example_name}) } },
-            .target = target,
-            .optimize = optimize,
-        });
         const example = b.addExecutable(.{
             .name = example_name,
-            .root_module = example_module,
+            .root_module = b.createModule(.{
+                .root_source_file = .{ .src_path = .{ .owner = b, .sub_path = b.fmt("examples/{s}.zig", .{example_name}) } },
+                .target = target,
+                .optimize = optimize,
+            }),
         });
         const install_example = b.addInstallArtifact(example, .{});
         const zbench_mod = b.addModule("zbench", .{