--- a/examples/bubble_sort.zig
+++ b/examples/bubble_sort.zig
@@ -20,7 +20,7 @@ fn myBenchmark(_: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    var stdout = std.fs.File.stdout().writer(&.{});
+    var stdout = std.fs.File.stdout().writerStreaming(&.{});
     var writer = &stdout.interface;
 
     var bench = zbench.Benchmark.init(std.heap.page_allocator, .{});