--- a/examples/bubble_sort_hooks.zig
+++ b/examples/bubble_sort_hooks.zig
@@ -49,7 +49,7 @@ fn afterAll() void {
 }
 
 pub fn main() !void {
-    var stdout = std.fs.File.stdout().writer(&.{});
+    var stdout = std.fs.File.stdout().writerStreaming(&.{});
     var writer = &stdout.interface;
 
     var bench = zbench.Benchmark.init(gpa.allocator(), .{});