--- a/examples/bubble_sort_hooks.zig
+++ b/examples/bubble_sort_hooks.zig
@@ -49,7 +49,7 @@ fn afterAll() void {
 }
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    const stdout = std.fs.File.stdout().deprecatedWriter();
 
     var bench = zbench.Benchmark.init(gpa.allocator(), .{});
     defer {