--- a/examples/hooks.zig
+++ b/examples/hooks.zig
@@ -18,7 +18,7 @@ fn myBenchmark(_: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    const stdout = std.fs.File.stdout().deprecatedWriter();
     var bench = zbench.Benchmark.init(std.heap.page_allocator, .{});
     defer bench.deinit();