--- a/examples/hooks.zig
+++ b/examples/hooks.zig
@@ -18,7 +18,7 @@ fn myBenchmark(_: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    var stdout = std.fs.File.stdout().writer(&.{});
+    var stdout = std.fs.File.stdout().writerStreaming(&.{});
     var writer = &stdout.interface;
 
     var bench = zbench.Benchmark.init(std.heap.page_allocator, .{});