--- a/examples/json.zig
+++ b/examples/json.zig
@@ -13,7 +13,7 @@ fn myBenchmark(alloc: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    var stdout = std.fs.File.stdout().writer(&.{});
+    var stdout = std.fs.File.stdout().writerStreaming(&.{});
     var writer = &stdout.interface;
 
     var bench = zbench.Benchmark.init(gpa.allocator(), .{});