--- a/examples/json.zig
+++ b/examples/json.zig
@@ -13,7 +13,7 @@ fn myBenchmark(alloc: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    const stdout = std.fs.File.stdout().deprecatedWriter();
 
     var bench = zbench.Benchmark.init(gpa.allocator(), .{});
     defer {