--- a/examples/memory_tracking.zig
+++ b/examples/memory_tracking.zig
@@ -10,7 +10,7 @@ fn myBenchmark(allocator: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    const stdout = std.fs.File.stdout().deprecatedWriter();
     var bench = zbench.Benchmark.init(gpa.allocator(), .{
         .iterations = 64,
     });