--- a/examples/memory_tracking.zig
+++ b/examples/memory_tracking.zig
@@ -10,7 +10,9 @@ fn myBenchmark(allocator: std.mem.Allocator) void {
 }
 
 pub fn main() !void {
-    const stdout = std.fs.File.stdout().deprecatedWriter();
+    var stdout = std.fs.File.stdout().writer(&.{});
+    var writer = &stdout.interface;
+
     var bench = zbench.Benchmark.init(gpa.allocator(), .{
         .iterations = 64,
     });
@@ -25,6 +27,6 @@ pub fn main() !void {
         .track_allocations = true,
     });
 
-    try stdout.writeAll("\n");
-    try bench.run(stdout);
+    try writer.writeAll("\n");
+    try bench.run(writer);
 }