--- a/examples/parameterised.zig
+++ b/examples/parameterised.zig
@@ -18,7 +18,7 @@ const MyBenchmark = struct {
 };
 
 pub fn main() !void {
-    const stdout = std.io.getStdOut().writer();
+    const stdout = std.fs.File.stdout().deprecatedWriter();
     var bench = zbench.Benchmark.init(std.heap.page_allocator, .{});
     defer bench.deinit();