--- a/examples/parameterised.zig
+++ b/examples/parameterised.zig
@@ -18,7 +18,7 @@ const MyBenchmark = struct {
 };
 
 pub fn main() !void {
-    var stdout = std.fs.File.stdout().writer(&.{});
+    var stdout = std.fs.File.stdout().writerStreaming(&.{});
     var writer = &stdout.interface;
 
     var bench = zbench.Benchmark.init(std.heap.page_allocator, .{});