This was merged recently (https://github.com/ziglang/zig/pull/23096) and requires some build.zig changes.