@FObersteiner 2025-07-27
@mrjbq7 I think this needs more changes, implementing the `std.Io` interface e.g. for printing to stdout etc. I'm not sure however if it would be wise to try to finish this yet; the whole "writergate" might not have evolved completely yet. @hendriknielaender do you have any insights from other Zig projects? I'm not doing much in Zig these days...