@mrjbq7 2025-07-27
The std.Io interface is not likely to be 0.15.0, but rather the next dev cycle.