@FObersteiner 2025-07-27
> The std.Io interface is not likely to be 0.15.0, but rather the next dev cycle.

got any refs for this? This not a critique btw. but an honest question - currently there seem to be a lot of breaking changes coming in the master branch of Zig and I'm unsure how to deal with it.

Anyhow, my point was that to make this PR complete, I think it should make the full library work with 0.15.0-dev, not just the build.zig