@mrjbq7 2025-07-27
Okay, I made the subsequent changes to update fully to the new ``std.io.Writer`` api.

I can't predict the future, but I would say they had been targeting an *early August* release of 0.15.0 and the ``std.Io`` interface changes are still in a branch:

https://github.com/ziglang/zig/tree/async-await-demo

In any event, this works cleanly with latest zig master nightly: 0.15.0-dev.1232+869ef0060

I tested by building and running all the examples.