@mrjbq7 2025-07-27
Currently in the examples, I do not provide a buffer for ``stdout``, you can see

```zig
var stdout = std.fs.File.stdout().writer(&.{});
var writer = &stdout.interface;
```

vs something like

```zig
var buffer: [1024]u8 = undefined;
var stdout = std.fs.File.stdout().writer(&buffer);
var writer = &stdout.interface;
```

Nothing required it for the examples, but if you do want that don't forget to

```zig
defer writer.flush() catch {};
```

In any event, the *upgrade guide* for your users now that it uses the new ``std.io.Writer`` type can include taking advantage of the fact that the Zig 0.15.0 will provide ``adaptToNewApi()`` functions on writers.  For example, using the ``stdout.deprecatedWriter()`` as an example:

```zig
var stdout = std.fs.File.stdout().deprecatedWriter();
var adapted = stdout.adaptToNewApi();
var writer = &adapted.new_interface;
```