@mrjbq7 2025-07-29
```
$ ./zig-out/bin/memory_comparison
Memory Usage Comparison: System Info Retrieval
==============================================


Stack Usage Comparison:
======================

benchmark              runs     total time     time/run (avg ± σ)    (min ... max)                p75        p99        p995
-----------------------------------------------------------------------------------------------------------------------------
Old Implementation (su 100      436.82ms       4.368ms ± 790.073us   (3.977ms ... 9.542ms)        4.228ms    9.542ms    9.542ms
Old Implementation (su [MEMORY]                5.394KiB ± 0B         (5.394KiB ... 5.394KiB)      5.394KiB   5.394KiB   5.394KiB
New Implementation (sy 100      253.583us      2.535us ± 402ns       (1.958us ... 4.25us)         2.792us    4.25us     4.25us
New Implementation (sy [MEMORY]                0B ± 0B               (0B ... 0B)                  0B         0B         0B
Old Stack Usage (24KB  100      2.042us        20ns ± 76ns           (0ns ... 750ns)              41ns       750ns      750ns
New Stack Usage (128B  100      1.666us        16ns ± 20ns           (0ns ... 42ns)               42ns       42ns       42ns
```